Install Tailwind, Storybook, and Redux Toolkit
Installed Tailwind for its utility-first approach to styling. Storybook will provide a way for the frontend to relay the look of various components to the other team members in a more organized, isolated environment. I am rather-safely assuming we will need state management of sorts, and Redux Toolkit is the go-to standard. Some minor configuration was also made to get these installed packages working appropriately.