Install react-router-dom, create routes and pages
As we need the ability to navigate across various pages in the site, we install react-router-dom -- the standard way to handle routing in React. Created several Page components that simply render a header with a description of what the page will eventually contain. We then create Routes to render these Page components at the appropriate URLs.

Note how I created an 'index.tsx' file that simply imports all the Page components, and exports them all as an object. I find this to be a commonly-used way to handle related components. As a result, in App.tsx we are able to import all the Page components we need with a single import statement!