Create NavButton, along with its Story
Created a NavButton -- a re-usable navigatonal button that receives the text for the button and a path to navigate to when the button is pressed. Although this is simply a wrapper for a regular HTML button with some basic Tailwind classes assigned, along with a react-router-dom Link, we can use this to have consistency among buttons on the site. It's also a good introduction component to test if Tailwind and Storbyook are working properly.