Begin mock-fetching and styling of Posts
Using the jsonplaceholder website, I fetch a small list of Posts and render them in a very rudimentary way to the page. Clicking a post will take you to the detail page for that Post, which includes any Comments found on it. Again, ditched the TSX extension in favor of non-TypeScript logic.