Add Post fetching from db, improve styling
Gave each post a background color that closer matches that of Reddit. In general, altered text colors, weights, and spacing to be more aligned with Reddit's. Various elements in a Post also direct the user to an appropriate page when clicked.

More importantly, the first Post listed on the page is being fetched from a temporary backend!