Lots of styling tweaks! More sorting! Voting!
A lot of changes here. Changes to the Posts screen to make it match the Reddit design more closely. Added up and down voting on a Post Preview. Added a chat icon. Added ability to sort by upvotes.

On the Post screen, added a textarea to enter your own post! It gets saved to the database. You can also up and downvote Comments.